requirements.txt
crewai>=0.11.0
langchain>=0.1.0
streamlit>=1.29.0
python-dotenv>=1.0.0
openai>=1.7.1
pandas>=2.0.0
duckduckgo-search
unstructured
tabulate
langchain_openai>=0.0.2
